Added working mpu6000 driver for cc3d flight controller

+++ arch/types.h
@@ -1,6 +1,8 @@
 #pragma once
 
+#include <inttypes.h>
+
 typedef uint16_t gpio_pin_t;
 
-typedef int32_t timestamp_t;
+typedef int64_t timestamp_t;
 typedef timestamp_t timestamp_t;
